lz4_flex::block

Function decompress_with_dict

pub fn decompress_with_dict(
    input: &[u8],
    min_uncompressed_size: usize,
    ext_dict: &[u8],
) -> Result<Vec<u8>, DecompressError>

Decompress all bytes of input into a new vec. The passed parameter min_uncompressed_size needs to be equal or larger than the uncompressed size.

Panics

May panic if the parameter min_uncompressed_size is smaller than the uncompressed data.