Struct lz4_flex::frame::FrameInfo

pub struct FrameInfo {
    pub content_size: Option<u64>,
    pub block_size: BlockSize,
    pub block_mode: BlockMode,
    pub block_checksums: bool,
    pub content_checksum: bool,
    pub legacy_frame: bool,
    /* private fields */
}

The metadata for de/compressing with lz4 frame format.

Fields
content_size: Option<u64>

If set, includes the total uncompressed size of data in the frame.

block_size: BlockSize

The maximum uncompressed size of each data block.

block_mode: BlockMode

The block mode.

block_checksums: bool

If set, includes a checksum for each data block in the frame.

content_checksum: bool

If set, includes a content checksum to verify that the full frame contents have been decoded correctly.

legacy_frame: bool

If set, use the legacy frame format

Implementations

impl FrameInfo

pub fn new() -> Self

Create a new FrameInfo.

pub fn content_size(self, content_size: Option<u64>) -> Self

Whether to include the total uncompressed size of data in the frame.

pub fn block_size(self, block_size: BlockSize) -> Self

The maximum uncompressed size of each data block.

pub fn block_mode(self, block_mode: BlockMode) -> Self

The block mode.

pub fn block_checksums(self, block_checksums: bool) -> Self

If set, includes a checksum for each data block in the frame.

pub fn content_checksum(self, content_checksum: bool) -> Self

If set, includes a content checksum to verify that the full frame contents have been decoded correctly.

pub fn legacy_frame(self, legacy_frame: bool) -> Self

If set, use the legacy frame format.