Enum lz4_flex::frame::Error

#[non_exhaustive]
pub enum Error {

    CompressionError(CompressError),
    DecompressionError(DecompressError),
    IoError(Error),
    UnsupportedBlocksize(u8),
    UnsupportedVersion(u8),
    WrongMagicNumber,
    ReservedBitsSet,
    InvalidBlockInfo,
    BlockTooBig,
    HeaderChecksumError,
    BlockChecksumError,
    ContentChecksumError,
    SkippableFrame(u32),
    DictionaryNotSupported,
    ContentLengthError {
        expected: u64,
        actual: u64,
    },

}

Errors that can occur when de/compressing lz4.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

CompressionError(CompressError)

Compression error.

DecompressionError(DecompressError)

Decompression error.

IoError(Error)

An io::Error was encountered.

UnsupportedBlocksize(u8)

Unsupported block size.

UnsupportedVersion(u8)

Unsupported frame version.

WrongMagicNumber

Wrong magic number for the LZ4 frame format.

ReservedBitsSet

Reserved bits set.

InvalidBlockInfo

Block header is malformed.

BlockTooBig

Read a block larger than specified in the Frame header.

HeaderChecksumError

The Frame header checksum doesnโt match.

BlockChecksumError

The block checksum doesnโt match.

ContentChecksumError

The content checksum doesnโt match.

SkippableFrame(u32)

Read an skippable frame. The caller may read the specified amount of bytes from the underlying io::Read.

DictionaryNotSupported

External dictionaries are not supported.

ContentLengthError

Content length differs.

Fields

expected: u64

Expected content length.

actual: u64

Actual content lenght.